requirements.txt
asgiref==3.3.1; python_version >= "3.6"
django==3.1.7; python_version >= "3.6"
djangorestframework==3.12.2; python_version >= "3.5"
pytz==2021.1; python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"